Fix stylesheet behavior
This way you can run it without a stylesheet. It'll be ugly, but it'll
run!

It also issues a bunch of warnings when you run it without a stylesheet:

WARNING: Ignored `text-rendering: optimizeLegibility` at 18:7, unknown property.
WARNING: Expected a media type, got '(max-width: 600px)'
WARNING: Invalid media type ' (max-width: 600px) ' the whole @media rule was ignored at 21:5.
WARNING: Ignored `overflow-x: auto` at 107:7, unknown property

I'm pretty sure what's going on here is that the default pandoc template
has a stylesheet that isn't fully weasyprint compatible. The real fix
here is to embed a stylesheet (and maybe even a full html template) in
the binary, and then use those instead of the defaults. That's a bunch
of work though, and I'm not inclined to do it until I actually need it!